adds heartbeat logic to base class
