+++ Common/rover_system.hpp
@@ -1,13 +1,16 @@
 #pragma once
 #include <string>
 
+#include "../Common/heartbeat.hpp"
+
 namespace sjsu::common
 {
 class RoverSystem
 {
  public:
-  struct RoverMissionControlData{
-    int is_operational = 0;
+  struct RoverMissionControlData
+  {
+    int is_operational  = 0;
     int heartbeat_count = 0;
   };
   /// Initialize all the motors and sensors that are used in the system
@@ -17,11 +20,38 @@ class RoverSystem
   /// Creates the GET request parameters that contain the current rover status
   virtual std::string GETParameters() = 0;
   /// Parses the JSON response retrieved from mission control
-  virtual void ParseJSONResponse(std::string &response) = 0;
+  virtual void ParseJSONResponse(std::string & response) = 0;
   /// Move the rover according to the data sent from mission control
   virtual void HandleRoverMovement() = 0;
-   /// Verifies that mission control is sending fresh commands to rover
-  virtual bool IsHeartbeatSynced() = 0;
 
+  bool IsHeartbeatSynced(int heartbeat_count)
+  {
+    if (heartbeat_count_ != heartbeat_count)
+    {
+      // TODO: Should throw error in an attempt to reconnect?
+      sjsu::LogError("Heartbeat out of sync - resetting!");
+      ResetHeartbeatCount();
+      return false;
+    }
+    return true;
+  }
+
+  int GetHeartbeatCount()
+  {
+    return heartbeat_count_;
+  }
+
+  void IncrementHeartbeatCount()
+  {
+    heartbeat_count_++;
+  }
+
+ private:
+  void ResetHeartbeatCount()
+  {
+    heartbeat_count_ = 0;
+  }
+
+  int heartbeat_count_ = 0;
 };
 }  // namespace sjsu::common
