+++ Drive/rover_drive_system.hpp
@@ -4,6 +4,7 @@
 
 #include "../Common/state_of_charge.hpp"
 #include "../Common/rover_system.hpp"
+#include "../Common/heartbeat.hpp"
 #include "../Common/esp.hpp"
 #include "wheel.hpp"
 
@@ -30,6 +31,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     int rotation_angle = 0;
     int speed          = 0;
   };
+
   RoverDriveSystem(Wheel & left_wheel, Wheel & right_wheel, Wheel & back_wheel)
       : left_wheel_(left_wheel),
         right_wheel_(right_wheel),
@@ -43,7 +45,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     back_wheel_.Initialize();
     SetSpinMode();
     sjsu::LogInfo("Drive system initialized!");
-  };
+  }
 
   /// Constructs parameters for an HTTP GET request
   /// @return ?heartbeat_count=0&is_operational=1&drive_mode=S ...
@@ -55,13 +57,13 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
         "?heartbeat_count=%d&is_operational=%d&drive_mode=%c&battery=%d"
         "&left_wheel_speed=%d&left_wheel_angle=%d&right_wheel_speed=%d&right_"
         "wheel_angle=%d&back_wheel_speed=%d&back_wheel_angle=%d",
-        heartbeat_count_, mc_data_.is_operational, current_mode_,
+        GetHeartbeatCount(), mc_data_.is_operational, current_mode_,
         state_of_charge_, left_wheel_.GetHubSpeed(),
         left_wheel_.GetSteerAngle(), right_wheel_.GetHubSpeed(),
         right_wheel_.GetSteerAngle(), back_wheel_.GetHubSpeed(),
         back_wheel_.GetSteerAngle());
     return request_parameter;
-  };
+  }
 
   /// Parses the GET requests response and updates the mission control variables
   void ParseJSONResponse(std::string & response) override
@@ -77,13 +79,13 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
                      kExpectedArguments);
       throw ParseError{};
     }
-  };
+  }
 
   /// Handles the rover movement depending on the mode.
   /// D = Drive, S = Spin, T = Translation, L/R/B = Left/Right/Back Wheel
   void HandleRoverMovement() override
   {
-    if (!IsHeartbeatSynced())
+    if (!IsHeartbeatSynced(mc_data_.heartbeat_count))
     {
       SetWheelSpeed(kZeroSpeed);
       return;
@@ -115,29 +117,6 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
         StopWheels();
         break;
     }
-  };
-
-  /// Checks if the mission control heartbeat matches rover heartbeat
-  bool IsHeartbeatSynced() override
-  {
-    if (GetHeartbeatCount() != mc_data_.heartbeat_count)
-    {
-      // TODO: Should throw error in an attempt to reconnect?
-      sjsu::LogError("Heartbeat out of sync - resetting!");
-      heartbeat_count_ = 0;
-      return false;
-    }
-    return true;
-  }
-
-  int GetHeartbeatCount()
-  {
-    return heartbeat_count_;
-  }
-
-  void IncrementHeartbeatCount()
-  {
-    heartbeat_count_++;
   }
 
   /// Checks if the rover is operational
@@ -198,7 +177,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     left_wheel_.SetHubSpeed(left_wheel_speed);
     right_wheel_.SetHubSpeed(right_wheel_speed);
     back_wheel_.SetHubSpeed(back_wheel_speed);
-  };
+  }
 
   /// Locks thread until all wheels are homed
   void HomeWheels()
@@ -227,7 +206,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
       sjsu::Delay(50ms);
     }
     sjsu::LogInfo("Wheels homed!");
-  };
+  }
 
   bool AllWheelsAreHomed()
   {
@@ -249,7 +228,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     right_wheel_.Print();
     back_wheel_.Print();
     printf("=========================\n");
-  };
+  }
 
  private:
   /// Stops the rover and sets a new mode.
@@ -267,7 +246,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
       case 'B': SetSingleWheelMode(); break;
       default: sjsu::LogError("Unable to set drive mode!");
     };
-  };
+  }
 
   // ======================
   // = DRIVE MODE SETTERS =
@@ -283,7 +262,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     right_wheel_.SetSteerAngle(right_wheel_angle);
     back_wheel_.SetSteerAngle(back_wheel_angle);
     current_mode_ = 'D';
-  };
+  }
 
   /// Aligns rover wheels perpendicular to their legs using homing slip ring
   void SetSpinMode()
@@ -295,7 +274,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     right_wheel_.SetSteerAngle(right_wheel_angle);
     back_wheel_.SetSteerAngle(back_wheel_angle);
     current_mode_ = 'S';
-  };
+  }
 
   /// Aligns rover wheel all in the same direction, facing towards the right
   void SetTranslationMode()
@@ -308,7 +287,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     right_wheel_.SetSteerAngle(right_wheel_angle);
     back_wheel_.SetSteerAngle(back_wheel_angle);
     current_mode_ = 'T';
-  };
+  }
 
   void SetSingleWheelMode()
   {
@@ -341,7 +320,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     }
     // TODO: Need logic for controling wheel speed for each wheel
     SetWheelSpeed(speed);
-  };
+  }
 
   /// Calculates outer wheel angle based off inner wheel angle
   float GetOutterWheelDriveAngle(float inner_wheel_angle)
@@ -367,7 +346,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
   void HandleSpinMode(float speed)
   {
     SetWheelSpeed(speed);
-  };
+  }
 
   /// Adjusts all the wheels by keeping them in parallel
   void HandleTranslationMode(float speed, float angle)
@@ -377,7 +356,7 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
     right_wheel_.SetSteerAngle(angle);
     back_wheel_.SetSteerAngle(angle);
     SetWheelSpeed(speed);
-  };
+  }
 
   /// Adjusts the hub speed and steer angle of the specified wheel
   void HandleSingularWheelMode(float speed, float angle)
@@ -397,9 +376,8 @@ class RoverDriveSystem : public sjsu::common::RoverSystem
         back_wheel_.SetHubSpeed(speed);
         break;
     }
-  };
+  }
 
-  int heartbeat_count_ = 0;
   int state_of_charge_ = 90;
   char current_mode_   = 'S';
