+++ Drive/source/main.cpp
@@ -41,7 +41,7 @@ int main(void)
                                  right_wheel_homing_pin);
   sjsu::drive::Wheel back_wheel("back", back_hub_motor, back_steer_motor,
                                 back_wheel_homing_pin);
-                                
+
   sjsu::drive::RoverDriveSystem drive(left_wheel, right_wheel, back_wheel);
 
   esp.Initialize();
